requirements.txt
asgiref==3.4.1
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.1
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.10.1
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
zipp==3.7.0
